Verificando código do aluno Emanuel Mesquita e inbcluindo alguns trechos de código da sala de aula

+++ src/main/java/com/emanuelcmesquita/dev/domain/Produto.java
@@ -1,17 +1,17 @@
 package com.emanuelcmesquita.dev.domain;
-
+//FIXME Incluir anotações do JPA , sem isso o projeto falha.
 public class Produto {
 
-    private int id;
+    private Long id;
     private String nome;
     private String descricao;
     private Double preco;
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
